+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Email verification feature
 - Change the User object to include timeJoined
 - Sends emails to our APIs only if not testing mode
-- Add useSuperTokensFromNextJs generic express middleware wrapper
+- Add superTokensNextWrapper generic express middleware wrapper
 
 ## [3.2.2] - 2020-12-18
 ### Fixed